pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "rnm-ivd"
version = "2.0.0"
description = "Regulatory Network Model for IVD NP Cell Signaling"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Sofia Tseranidou", email = "sofiatseranidou@upf.edu"},
    {name = "Maria Segarra-Queralt", email = "maria.segarra@upf.edu"},
    {name = "Francis Kiptengwer Chemorion", email = "francis.chemorion@upf.edu"},
    {name = "Zerihun G. Workineh", email = "zerihungetahun.workineh@upf.edu"},
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "openpyxl>=3.1",
    "python-libsbml>=5.20",
]

[project.optional-dependencies]
sim = ["libroadrunner>=2.7"]

[tool.setuptools.packages.find]
include = ["rnm*"]
